Restructuring of project folders
